requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
schedule>=1.2.0
streamlit-extras>=0.3.0
streamlit-option-menu>=0.3.0
